# accelerate jobmanager loop