--- a/src/ClusterManager/cluster_manager.py
+++ b/src/ClusterManager/cluster_manager.py
@@ -87,10 +87,18 @@ def Run(args):
 
     cwd = os.path.dirname(__file__)
     cmds = {
-        "job_manager1":
-        ["python", os.path.join(cwd, "job_manager.py"), "--port", str(args.j1), "--updateblock", "1"],
-        "job_manager2":
-        ["python", os.path.join(cwd, "job_manager.py"), "--port", str(args.j2), "--updateblock", "2"],        
+        "job_manager_killing,pausing,unapproved":
+        ["python", os.path.join(cwd, "job_manager.py"), "--port", str(args.j1),
+            "--status", "killing,pausing,unapproved"],
+        "job_manager_running":
+        ["python", os.path.join(cwd, "job_manager.py"), "--port", str(args.j2),
+            "--status", "running"],
+        "job_manager_scheduling":
+        ["python", os.path.join(cwd, "job_manager.py"), "--port", str(args.j3),
+            "--status", "scheduling"],
+        "job_manager_queued":
+        ["python", os.path.join(cwd, "job_manager.py"), "--port", str(args.j4),
+            "--status", "queued"],
         "user_manager":
         ["python", os.path.join(cwd, "user_manager.py"), "--port", str(args.u)],
         "node_manager":
@@ -149,6 +157,8 @@ if __name__ == "__main__":
     parser.add_argument("--tictoc", help="how many seconds to wait until kill subprocess", type=int, default=600)
     parser.add_argument("-j1", help="port of job_manager", type=int, default=9200)
     parser.add_argument("-j2", help="port of job_manager", type=int, default=9206)
+    parser.add_argument("-j3", help="port of job_manager", type=int, default=9207)
+    parser.add_argument("-j4", help="port of job_manager", type=int, default=9208)
     parser.add_argument("-u", help="port of user_manager", type=int, default=9201)
     parser.add_argument("-n", help="port of node_manager", type=int, default=9202)
     parser.add_argument("-l", help="port of joblog_manager", type=int, default=9203)